feat: refactor header components and styles for improved structure and consistency
