+++ mobile/app/(auth)/login.tsx
@@ -15,7 +15,7 @@ import {
 import { useRouter } from 'expo-router';
 import { useAuth } from '../../context/AuthContext';
 import { styles } from '../../styles/auth/loginSytles';
-import Header from '../../components/Header';
+import Header from '../../components/shared/AuthHeader';
 import { LinearGradient } from 'expo-linear-gradient';
 import { landingPageColors } from '../../constants/IndexColors';
