+++ mobile/app/(auth)/plans.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { View, Text, StyleSheet, ScrollView } from 'react-native';
 import { Link } from 'expo-router';
 import { Pressable, TouchableOpacity } from 'react-native-gesture-handler';
-import Header from '../../components/Header';
+import Header from '../../components/shared/AuthHeader';
 import { styles } from '../../styles/auth/plansSytles';
 import { LinearGradient } from 'expo-linear-gradient';
 import { landingPageColors } from '../../constants/IndexColors';
