+++ mobile/app/(tabs)/stock.tsx
@@ -5,6 +5,7 @@ import type { Product } from '../../services/stockService';
 import ProductList from '../../components/stock/ProductList';
 import AddProductModal from '../../components/stock/AddProductModal';
 import { DashboardColors } from '@/constants/DashboardColors';
+import { styles } from '../../styles/stock/Stock';
 
 export default function StockScreen() {
     const [products, setProducts] = useState<Product[]>([]);
@@ -68,36 +69,3 @@ export default function StockScreen() {
     );
 }
 
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        backgroundColor: DashboardColors.background,
-    },
-    center: {
-        justifyContent: 'center',
-        alignItems: 'center',
-    },
-    pageHeader: {
-        flexDirection: 'row',
-        justifyContent: 'space-between',
-        alignItems: 'center',
-        paddingHorizontal: 20,
-        paddingTop: 10,
-        paddingBottom: 20,
-    },
-    title: {
-        fontSize: 24,
-        fontWeight: 'bold',
-        color: DashboardColors.darkText,
-    },
-    addButton: {
-        backgroundColor: DashboardColors.headerBlue,
-        paddingHorizontal: 16,
-        paddingVertical: 8,
-        borderRadius: 20,
-    },
-    addButtonText: {
-        color: 'white',
-        fontWeight: 'bold',
-    },
-});
