+++ mobile/styles/shared/AppHeaderStyles.ts
