+++ mobile/styles/shared/AuthHeaderStyles.ts
@@ -1,0 +1,15 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+  header: {
+    backgroundColor: 'rgba(255, 255, 255, 0.95)',
+    padding: 16,
+    alignItems: 'center',
+    borderBottomLeftRadius: 20,
+    borderBottomRightRadius: 20,
+  },
+  logo: {
+    width: 80,
+    height: 80,
+  },
+});
