+++ mobile/styles/stock/Stock.ts
@@ -1,0 +1,36 @@
+import { StyleSheet } from 'react-native';
+import { DashboardColors } from '@/constants/DashboardColors';
+
+export const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        backgroundColor: DashboardColors.background,
+    },
+    center: {
+        justifyContent: 'center',
+        alignItems: 'center',
+    },
+    pageHeader: {
+        flexDirection: 'row',
+        justifyContent: 'space-between',
+        alignItems: 'center',
+        paddingHorizontal: 20,
+        paddingTop: 10,
+        paddingBottom: 20,
+    },
+    title: {
+        fontSize: 24,
+        fontWeight: 'bold',
+        color: DashboardColors.darkText,
+    },
+    addButton: {
+        backgroundColor: DashboardColors.headerBlue,
+        paddingHorizontal: 16,
+        paddingVertical: 8,
+        borderRadius: 20,
+    },
+    addButtonText: {
+        color: 'white',
+        fontWeight: 'bold',
+    },
+});
